Store RIP on the stack when entering an interrupt
Storing it in RID fails for two reasons: 1) it corrupts RID, which
is critical if the interrupt occured in between instructions
which share RID (e.g. CMP expands to MOV xxx rid, SUB xxx rid)
and 2) (the original reason I thought of this) if an NMI occurs
immediately after a normal interrupt, then the original RIP will be
lost as RID is written again now with the address of the interrupt
handler.